WordCamp Calendar: Explicitly send a full CRLF.
RFC 2445 requires a full CRLF, but previously we were just sending \\n. That may not have mattered at all in the real world, but I've gotten reproducable reports of Google Calendar failing to import any of the events from the feed. This didn't end up fixing the problem, but I'm committing it anyway since the work is already done and it brings the output in line with the RFC.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@480 74240141-8908-4e6f-9713-ba540dce6ec7