WordCamp Organizer Reminders: Get address from $_POST instead of database.
If the post is a draft and the email field is empty, then the admin enters the email and saves sets it to pending
in a single step, the previous code would have pulled an empty string from the database instead of the new address
from $_POST, and the e-mail would not have been sent.

Pulling from $_POST instead of the database ensures that the most recent value will always be retrieved.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@760 74240141-8908-4e6f-9713-ba540dce6ec7