WordCamp Payments Dashboard: Reorganize the contents of the tabs.
Overview contained all statuses, which will quickly become cluttered with paid requests, and therefore be useless for seeing the most relevant requests quickly. Overdue is now the default tab, since it's the most urgent. Paid requests have been moved to their own tab, so that they're accessible, but out of the way.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@1222 74240141-8908-4e6f-9713-ba540dce6ec7