WordCamp Budgets: Enable (but hide) Sponsor Invoice for unproxied users.
This is a safer way to make the new module available for beta testing on production, while still preventing regular users from accessing it before it's ready. The previous approach led to several instances of PHP errors that didn't show up during pre-commit testing, because I was always simulated as being proxied on my sandbox.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@2613 74240141-8908-4e6f-9713-ba540dce6ec7