WordCamp Budgets: Remove redundant country fields.
ISO 3166-compliant country fields were added in r2420, and these old ones should have been removed then.

We don't need to worry about back-compat because there aren't any current requests that are missing the new fields, and the data for the old fields is still in the database if we ever need it.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@2701 74240141-8908-4e6f-9713-ba540dce6ec7