WordCamp.org: Create a new site a bit later.
The `wcpt_metabox_save` action is a bit misleading because
it runs multiple times, so the multi-sponsor region is saved
after the WordCamp URL is saved, causing our "create site"
functionality to fail the first time.

This commit creates a new `wcpt_metabox_save_done` action which
is called when all metabox fields are saved. Moves the site
creation process to that new action.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@2962 74240141-8908-4e6f-9713-ba540dce6ec7