WordCamp Coming Soon Page: Fix broken menu link.
This bug didn't exist in the original patch, but was introduced during the review phase, because I didn't connect the dots between `register_settings_pages()` and `get_customizer_section_url()`, and thought it'd be better to use an absolute URL.

The bug didn't show up during local testing because of differences between the nginx configuration on my sandbox and the config on production, and I failed to test production after deploying.

This commit reverts it back to a relative URL so that it will work again, and also moves it inside `register_settings_pages()` to avoid confusing it with a complete URL that would be safe to use in other contexts.

See #334
Props ryelle

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@3835 74240141-8908-4e6f-9713-ba540dce6ec7