WordCamp: Move camptix-network-tools plugin to Meta repo
This removes the external that links to plugins.svn.wordpress.org and 
adds the files from the plugin directly to meta.svn.wordpress.org.

CampTix Network Tools is no longer maintained for use outside of WordCamp.org,
so this makes the Meta repo the canonical open source of the plugin.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins@7217 74240141-8908-4e6f-9713-ba540dce6ec7